requirements.txt
python-telegram-bot[job-queue]==20.7
google-api-python-client==2.136.0
google-auth==2.32.0
